Move the location selection and date selection into own modules
